+++ CHANGES.md
@@ -8,6 +8,8 @@ Version 0.3.0
 
 To be released.
 
+ -  Added *combinators.ts* module.
+ -  Added `tee()` function.
  -  Added *drop.ts* module.
  -  Added `drop()` function.
  -  Added `dropWhile()` function.